Use u/s32 for intptr_t

+++ include/stdint.h
@@ -3,8 +3,8 @@
 
 #include <ultra64.h>
 
-typedef long intptr_t;
-typedef unsigned long uintptr_t;
+typedef s32 intptr_t;
+typedef u32 uintptr_t;
 
 #define INT8_MIN    (-0x80)
 #define INT16_MIN   (-0x8000)
